obj: implement allocation alignment

+++ src/tools/pmempool/info_obj.c
@@ -677,7 +677,7 @@ info_obj_chunk(struct pmem_info *pip, uint64_t c, uint64_t z,
 
 		struct alloc_class *aclass = alloc_class_by_run(
 			pip->obj.alloc_classes,
-			run->block_size, m.header_type, m.size_idx);
+			run->block_size, (uint16_t)m.header_type, m.size_idx);
 		if (aclass) {
 			outv_field(v, "Block size", "%s",
 					out_get_size_str(run->block_size,
